#1917# PHP 5.0.5 changes into BRANCH-2_0

+++ classes/mail/MailTemplate.inc.php
@@ -82,14 +82,14 @@ class MailTemplate extends Mail {
 			$this->setBody(Request::getUserVar('body'));
 			$this->enabled = true;
 
-			if (is_array(Request::getUserVar('to'))) {
-				$this->setRecipients($this->processAddresses ($this->getRecipients(), Request::getUserVar('to')));
+			if (is_array($toEmails = Request::getUserVar('to'))) {
+				$this->setRecipients($this->processAddresses ($this->getRecipients(), $toEmails));
 			}
-			if (is_array(Request::getUserVar('cc'))) {
-				$this->setCcs($this->processAddresses ($this->getCcs(), Request::getUserVar('cc')));
+			if (is_array($ccEmails = Request::getUserVar('cc'))) {
+				$this->setCcs($this->processAddresses ($this->getCcs(), $ccEmails));
 			}
-			if (is_array(Request::getUserVar('bcc'))) {
-				$this->setBccs($this->processAddresses ($this->getBccs(), Request::getUserVar('bcc')));
+			if (is_array($bccEmails = Request::getUserVar('bcc'))) {
+				$this->setBccs($this->processAddresses ($this->getBccs(), $bccEmails));
 			}
 		}
 
@@ -375,7 +375,8 @@ class MailTemplate extends Mail {
 		import('file.TemporaryFileManager');
 		$temporaryFileManager = new TemporaryFileManager();
 
-		foreach (Request::getUserVar('persistAttachments') as $fileId) {
+		$persistAttachments = Request::getUserVar('persistAttachments');
+		if (is_array($persistAttachments)) foreach ($persistAttachments as $fileId) {
 			$temporaryFile = $temporaryFileManager->getFile($fileId, $userId);
 			if (!empty($temporaryFile)) {
 				$temporaryFileManager->deleteFile($temporaryFile->getFileId(), $userId);
